Move frame group into separate file
